Hi 👋 

First off, thanks for making this gem! I love it ❤️ 

When I was looking at the render implementation I had a few ideas for things that might make it a little faster. I knew I would need to add a benchmark to measure any results though so I added that first. While I was setting that up I discovered that there was an implicit dependency on active_support with the use of `present?`, which is not a Ruby core method. I removed that by expanding those uses out into what `present?` would have done, which is a falsey check, and then an empty check.


<details>
<summary>Then I added the benchmark and took initial measurements</summary>

```
Warming up --------------------------------------
rendering only defaults
                       129.242k i/100ms
rendering with overrides
                       107.147k i/100ms
Calculating -------------------------------------
rendering only defaults
                          1.278M (± 2.4%) i/s -      6.462M in   5.060689s
rendering with overrides
                          1.051M (± 2.6%) i/s -      5.250M in   5.000286s
```

</details>

The main performance gains came from 3 changes as far as I could tell:
* Merging `overrides` (née `settings`) onto defaults, and iterating that result once. Instead of iterating both of them individually and keeping track of which settings were set
* Using the instance variables directly instead of accessing them through the `attr_reader`s
* Removing the conditional checks before appending to `result` and instead using a single `compact!` at the end

I tried out a few other things I could think of, but they only made the performance worse.

<details>
<summary>This was my final benchmark after these performance improvements were made</summary>

```
Perf
Warming up --------------------------------------
rendering only defaults
                       186.738k i/100ms
rendering with overrides
                       162.322k i/100ms
Calculating -------------------------------------
rendering only defaults
                          1.851M (± 2.3%) i/s -      9.337M in   5.045712s
rendering with overrides
                          1.610M (± 2.5%) i/s -      8.116M in   5.045476s
```

</details>

According to my napkin math that's about a 55% improvement on the "rendering with overrides" case, and a 44% improvement on the "rendering only defaults" case.