@adrianthedev 2022-12-20
🙌 Just a perfect PR!

I pretty much made this late one night and never thought about performance. Your updates make perfect sense.

Thanks for contributing!